feat: watcher 과제 build, run log 관련 api 중계 설정 & 과제,강의 일치 여부 검증 추가

+++ src/main/kotlin/org/jbnu/jdevops/jcodeportallogin/dto/watcher/AssingmentTotalGraphData.kt
@@ -1,7 +1,7 @@
 package org.jbnu.jdevops.jcodeportallogin.dto.watcher
 
 data class AssingmentTotalGraphData(
-    val student_num: Long,
+    val student_num: Int,
     val size_change: Long
 )
